Update mean confidence param to nullable float

+++ Is This Vegan - API/Is This Vegan (Net)/Backend/Interfaces/IPipeline.cs
@@ -10,6 +10,6 @@ namespace Is_This_Vegan__Net_.Backend.Interfaces
 {
     public interface IPipeline
     {
-        PipelineResultModel Execute<T>(ref T input, DataCleanEnum? type, double? meanConfidence);
+        PipelineResultModel Execute<T>(ref T input, DataCleanEnum? type, float? meanConfidence);
     }
 }
